% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{calc_contribution}
\alias{calc_contribution}
\title{Calculate score contributions by phosphorylation site}
\usage{
calc_contribution(matched_data)
}
\arguments{
\item{matched_data}{Input}
}
\description{
Calculate score contributions by phosphorylation site
}
\keyword{internal}
