% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_calcs.R
\name{getsubs}
\alias{getsubs}
\title{Extract summary table with pertinent columns related to included substrates}
\usage{
getsubs(matched_data)
}
\arguments{
\item{matched_data}{Input with calculated contributions}
}
\description{
Extract summary table with pertinent columns related to included substrates
}
\keyword{internal}
