% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{KSEAfilter}
\alias{KSEAfilter}
\title{Filter matched data to remove positive IDs from KSEA}
\usage{
KSEAfilter(matched_data, kseadb, reverse = F)
}
\arguments{
\item{matched_data}{Results of get_matched_data function}

\item{kseadb}{KSEA database containing substrate gene name "SUB_GENE" and phosphorylated residue
"SUB_MOD_RSD" in standard form (ie. T302).}
}
\description{
Filter matched data to remove positive IDs from KSEA
}
\keyword{internal}
