% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{mk_runlabel}
\alias{mk_runlabel}
\title{mk_runlabel()}
\usage{
mk_runlabel(parentdir = getwd(), customsuffix)
}
\arguments{
\item{parentdir}{parent directory}

\item{customsuffix}{additional suffix to run identifier}
}
\description{
Utility function for generating new identifiers for each run, labeled by time run was initiated and
custom suffix
}
\keyword{internal}
