% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_calcs.R
\name{perc.permutation}
\alias{perc.permutation}
\title{Get percentile ranks across permutations}
\usage{
perc.permutation(results, permutations)
}
\arguments{
\item{results}{Results of kinase scoring}

\item{permutations}{Results of permutations}
}
\description{
Get percentile ranks across permutations
}
\keyword{internal}
