% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_calcs.R
\name{perc.rank}
\alias{perc.rank}
\title{Obtain percentile rank comparing a single value to set}
\usage{
perc.rank(set, value)
}
\arguments{
\item{set}{Set of values to which given value will be compared}

\item{value}{Value for which percentile score will be calculated}
}
\description{
Obtain percentile rank comparing a single value to set
}
\keyword{internal}
