% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{getscores}
\alias{getscores}
\title{Sum score contributions for each kinase across all phosphopeptides}
\usage{
getscores(matched_data)
}
\arguments{
\item{matched_data}{Input with calculated contributions}
}
\description{
Sum score contributions for each kinase across all phosphopeptides
}
\keyword{internal}
