% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{KSEAdb}
\alias{KSEAdb}
\title{KSEAdb}
\format{An object of class \code{data.frame} with 240749 rows and 14 columns.}
\source{
\url{https://www.phosphosite.org/staticDownloads.action}
}
\usage{
KSEAdb
}
\description{
A data table containing all known kinase-substrate links known in PhosphoSitePlus.
}
\keyword{datasets}
