% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{NetworKINPred_db}
\alias{NetworKINPred_db}
\title{NetworKINPred_db}
\format{An object of class \code{data.frame} with 450418 rows and 4 columns.}
\source{
\url{http://networkin.info/download.shtml}
}
\usage{
NetworKINPred_db
}
\description{
A data table containing all precalculated NetworKIN predictions performed on known ensembl sequences.
}
\references{
Horn et al., KinomeXplorer: an integrated platform for kinome biology studies. Nature Methods 2014 Jun;11(6):603–4.
}
\keyword{datasets}
