% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_calcs.R
\name{permtest}
\alias{permtest}
\title{Perform permutation test}
\usage{
permtest(matched_data, perms = 1000, seed = 1)
}
\arguments{
\item{matched_data}{Input with calculated contributions}

\item{perms}{Number of permutations to run, default = 1000}
}
\description{
Returns a table that has permuted the relationship between phosphopeptides and summary statistics
(ie. fold change, t-score)
}
\keyword{internal}
