% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_and_output_b.R
\name{results_write}
\alias{results_write}
\title{Output writing of pKSEA compare() results}
\usage{
results_write(full_ksea.results, outputpath, outputname, singlefolder = NULL)
}
\arguments{
\item{full_ksea.results}{results from compare() including full and optional KSEA excluded and exclusive results}

\item{outputpath}{parent directory for output}

\item{outputname}{file name of output}

\item{singlefolder}{if desired, name of output folder within parent directory. Default is separate folders
for each compare() run}
}
\description{
Output only:
uses results from compare(), outputs up to three files labeled full.csv and no_ksea.csv and ksea_only.csv
appended to an output name (KSEA-filtered results only if KSEA database was provided to compare()).
}
\keyword{internal}
